airflow.requirements.txt
# Add any additional Python dependencies below:
cx-Oracle==7.3.0 
d6tstack 
d6tstack[psql] 
google-api-python-client  
google-auth-httplib2  
google-auth-oauthlib  
google-cloud-storage  
gcsfs  
xlrd  
openpyxl  
unidecode  
sendgrid  
pep8 
pylint 
pandarallel 
great_expectations==0.13.14 
airflow-provider-great-expectations 
emails 
apache-airflow-backport-providers-amazon
striprtf==0.0.14
pika
objdict
GitPython